♻️ split github infrastructure
