+++ event-listeners/Cargo.toml
@@ -22,6 +22,9 @@ uuid = {version = "0.8.2", default_features = false, features = [
   "serde",
 ]}
 
+# Github
+octocrab = "0.16.0"
+
 # Http client
 reqwest = {version = "0.11", default-features = false}
 url = "2.2.2"
