+++ marketplace-domain/src/services/github_client.rs
@@ -1,6 +1,3 @@
-use crate::{
-	GithubIssue, GithubIssueNumber, GithubProjectId, GithubRepo, GithubUser, GithubUserId,
-};
 use async_trait::async_trait;
 use mockall::automock;
 use thiserror::Error;
@@ -11,19 +8,10 @@ pub enum Error {
 	Infrastructure(anyhow::Error),
 }
 
+pub type UserId = u64;
+
 #[automock]
 #[async_trait]
 pub trait GithubClient: Send + Sync {
-	async fn find_issue_by_id(
-		&self,
-		project_id: GithubProjectId,
-		issue_number: GithubIssueNumber,
-	) -> Result<GithubIssue, Error>;
-
-	async fn find_repository_by_id(&self, project_id: GithubProjectId)
-	-> Result<GithubRepo, Error>;
-
-	async fn find_user_by_id(&self, user_id: GithubUserId) -> Result<GithubUser, Error>;
-
-	async fn authenticate_user(&self, authorization_code: String) -> Result<GithubUserId, Error>;
+	async fn authenticate_user(&self, authorization_code: String) -> Result<UserId, Error>;
 }
